Caught an exception in the framerate call as ZeroDivionError wasn't caught and could occur if the metadata says the framerate is '0/0'
Added a new class called FFContainer that can be used to utilize to capture info about the container itself versus the stream.

Added a function to FFProbe to break out the calling of ffprobe so future development/changes could call ffprobe with new arguments in certain cases.

I also added type hinting to help with development of this and usage of this.